back/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==5.0.0
boto3==1.42.68
botocore==1.42.68
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
coverage==7.13.5
cryptography==46.0.2
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.119.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jmespath==1.1.0
packaging==26.0
passlib==1.7.4
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==12.0.0
pluggy==1.6.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
psycopg[binary]
Pygments==2.20.0
pypdfium2==4.30.0
pytest==9.0.3
pytest-cov==7.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.20
requests==2.32.5
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
